Move Rand tests to unit tests.
