--- test/RandTest/include/Resources.h
@@ -1,12 +1,0 @@
-#pragma once
-#include "cinder/CinderResources.h"
-
-//#define RES_MY_RES			CINDER_RESOURCE( ../resources/, image_name.png, 128, IMAGE )
-
-
-
-
-
-
-
-
