--- test/RandTest/vc11/RandTest.sln
@@ -1,20 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Express 2012 for Windows Desktop
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RandTest", "RandTest.vcxproj", "{8318FAD8-797D-45CD-87EF-B214930C5FD5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Release|Win32 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8318FAD8-797D-45CD-87EF-B214930C5FD5}.Debug|Win32.ActiveCfg = Debug|Win32
-		{8318FAD8-797D-45CD-87EF-B214930C5FD5}.Debug|Win32.Build.0 = Debug|Win32
-		{8318FAD8-797D-45CD-87EF-B214930C5FD5}.Release|Win32.ActiveCfg = Release|Win32
-		{8318FAD8-797D-45CD-87EF-B214930C5FD5}.Release|Win32.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
