--- test/RandTest/vc11/Resources.rc
@@ -1,3 +1,0 @@
-#include "../include/Resources.h"
-
-1	ICON	"..\\..\\..\\samples\\data\\cinder_app_icon.ico"
