--- test/RandTest/xcode_ios/RandTest_Prefix.pch
@@ -1,17 +1,0 @@
-#if defined( __OBJC__ )
-	#import <UIKit/UIKit.h>
-	#import <Foundation/Foundation.h>
-#endif
-
-#if defined( __cplusplus )
-	#include "cinder/Cinder.h"
-	
-	#include "cinder/app/App.h"
-	
-	#include "cinder/gl/gl.h"
-	
-	#include "cinder/CinderMath.h"
-	#include "cinder/Matrix.h"
-	#include "cinder/Vector.h"
-	#include "cinder/Quaternion.h"
-#endif
